backend/requirements.txt
fastapi
uvicorn[standard]
eth-account
PyJWT
openai
python-dotenv
httpx
email-validator
psycopg[binary]
SQLAlchemy
